refactor: add useBookSearch hook
